feat(ovh-at-internet): remove lodash dependency
ref: MANAGER-16864
